requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.6.0
transformers>=4.30.0
wandb>=0.15.0
tqdm>=4.65.0
pillow>=9.0.0
numpy>=1.24.0
tensorboard==2.14.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.76
